refactor supabase code
