+++ src/supabase_sync.ts
@@ -1,0 +1,159 @@
+import { AuthResponse, createClient } from "@supabase/supabase-js";
+import { Note } from "main";
+import { FleetingNotesSettings } from "settings";
+import { decryptNote, encryptNote, throwError } from "utils";
+
+const supabase = createClient(
+  "https://yixcweyqwkqyvebpmdvr.supabase.co",
+  "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InlpeGN3ZXlxd2txeXZlYnBtZHZyIiwicm9sZSI6ImFub24iLCJpYXQiOjE2NjQ4MDMyMTgsImV4cCI6MTk4MDM3OTIxOH0.awfZKRuaLOPzniEJ2CIth8NWPYnelLfsWrMWH2Bz3w8"
+);
+
+interface SupabaseNote {
+	id: string;
+	title: string;
+	content: string;
+	source: string;
+	created_at: string;
+	modified_at: string;
+	deleted: boolean;
+	shared: boolean;
+	encrypted: boolean;
+	_partition: string;
+}
+
+class SupabaseSync {
+  settings: FleetingNotesSettings;
+  constructor(settings: FleetingNotesSettings) {
+    this.settings = settings
+  }
+
+  updateNotes = async (notes: any[]) => {
+    try {
+      let supabaseNotes: SupabaseNote[] = [];
+      let noteIds = new Set(notes.map((note) => note.id));
+      // get all fields of the note
+      const res = await supabase
+        .from("notes")
+        .select()
+        .in("_partition", [this.settings.firebaseId, this.settings.supabaseId])
+        .eq("deleted", false)
+
+      if (res.error) {
+        throwError(res.error, res.error.message);
+      }
+      supabaseNotes = res.data;
+
+      // only take notes that are modified after note from db & note exists
+      notes = notes.filter((note) => {
+        let supabaseNote = res.data.find(
+          (supabaseNote: any) => 
+          supabaseNote.id === note.id && noteIds.has(supabaseNote.id)
+        );
+        return (supabaseNote) ? true : false;
+      });
+
+      // merge possibly updated fields
+      notes = notes.map((note) => {
+        let supabaseNote = supabaseNotes?.find(
+          (supabaseNote: any) => supabaseNote.id === note.id
+        );
+        var newNote = {
+          ...supabaseNote,
+          title: note.title || supabaseNote.title,
+          content: note.content || supabaseNote.content,
+          source: note.source || supabaseNote.source,
+          modified_at: new Date().toISOString(),
+          deleted: note.deleted || supabaseNote.deleted,
+        };
+        return (supabaseNote.encrypted) ? encryptNote(newNote, this.settings.encryption_key) : newNote;
+      });
+
+      if (notes.length > 0) {
+        const { error } = await supabase
+          .from("notes")
+          .upsert(notes, {
+            onConflict: "id",
+          })
+        if (error) {
+          throwError(error, error.message);
+        }
+      }
+    } catch (e) {
+      throwError(
+        e,
+        "Failed to update notes in Fleeting Notes"
+      );
+    }
+  }
+
+  getAllNotes = async () => {
+    let notes: Note[] = [];
+    try {
+      if (!this.settings.firebaseId) {
+        throwError(
+          "Fleeting Notes Sync Failed - Please Log In",
+          "Fleeting Notes Sync Failed - Please Log In"
+        );
+      }
+      await supabase
+        .from("notes")
+        .select()
+        .filter("_partition", "in", `(${this.settings.firebaseId},${this.settings.supabaseId})`)
+        .filter("deleted", "eq", false)
+        .then((res) => {
+          if (res.error) {
+            throwError(res.error, res.error.message);
+          }
+          notes = Array.from(
+            res.data?.map((note: any) => decryptNote(note, this.settings.encryption_key)) || []
+          );
+          if (this.settings.notes_filter) {
+            notes = notes.filter(
+              (note) =>
+                note.title.includes(this.settings.notes_filter) ||
+                note.content.includes(this.settings.notes_filter)
+            );
+          }
+        });
+      return notes;
+    } catch (e) {
+      throwError(
+        e,
+        "Failed to get notes from Fleeting Notes - Check your credentials"
+      );
+    }
+    return notes;
+  }
+
+  // supabase auth stuff
+  static loginSupabase = async (
+    email: string,
+    password: string
+  ): Promise<AuthResponse> => {
+    try {
+      const supaRes: AuthResponse = await supabase.auth
+        .signInWithPassword({
+          email,
+          password,
+        });
+      if (supaRes.error) {
+        throwError(supaRes.error, supaRes.error.message);
+      }
+      return supaRes;
+    } catch (err) {
+      throwError(err, err.message);
+    }
+  };
+
+  static onAuthStateChange = async (callback: (event: string) => void) => {
+    // check user logged in
+    supabase.auth.getUser().then((v) => {
+      if (!v.data?.user) {
+        callback("SIGNED_OUT");
+      }
+    });
+    return supabase.auth.onAuthStateChange(callback);
+  }
+}
+
+export default SupabaseSync;
