+++ src/utils.ts
@@ -1,14 +1,7 @@
 import { Note } from "./main";
 var CryptoJS = require("crypto-js");
-import { AuthResponse, createClient } from "@supabase/supabase-js";
 import { moment } from "obsidian"
 import { InputModal, Values, ModalInputField } from "./components/inputModal";
-
-// Create a single supabase client for interacting with your database
-const supabase = createClient(
-	"https://yixcweyqwkqyvebpmdvr.supabase.co",
-	"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InlpeGN3ZXlxd2txeXZlYnBtZHZyIiwicm9sZSI6ImFub24iLCJpYXQiOjE2NjQ4MDMyMTgsImV4cCI6MTk4MDM3OTIxOH0.awfZKRuaLOPzniEJ2CIth8NWPYnelLfsWrMWH2Bz3w8"
-);
 
 export function openInputModal(
 	title: string,
@@ -36,170 +29,6 @@ export function throwError(e: any, errMessage: string) {
 		throw errMessage;
 	}
 }
-
-export const loginSupabase = async (
-	email: string,
-	password: string
-): Promise<any> => {
-	try {
-		const supaRes: AuthResponse = await supabase.auth
-			.signInWithPassword({
-				email,
-				password,
-			})
-			.then((res) => {
-				return res;
-			});
-    if (supaRes.error) {
-      throwError(supaRes.error, supaRes.error.message);
-    }
-		return supaRes;
-	} catch (err) {
-		throwError(err, err.message);
-	}
-};
-
-export const onAuthStateChange = async (callback: (event: string) => void) => {
-  // check user logged in
-  supabase.auth.getUser().then((v) => {
-    if (!v.data?.user) {
-      callback("SIGNED_OUT");
-    }
-  });
-  return supabase.auth.onAuthStateChange(callback);
-}
-
-export const getAllNotesSupabase = async ({
-	firebaseId,
-	supabaseId,
-	key,
-	filterKey,
-}: {
-	firebaseId: string;
-	supabaseId: string;
-	key: string;
-	filterKey: string;
-}) => {
-	let notes: Note[] = [];
-
-	try {
-		if (!firebaseId) {
-			throwError(
-				"Fleeting Notes Sync Failed - Please Log In",
-				"Fleeting Notes Sync Failed - Please Log In"
-			);
-		}
-		await supabase
-			.from("notes")
-			.select()
-			.filter("_partition", "in", `(${firebaseId},${supabaseId})`)
-			.filter("deleted", "eq", false)
-			.then((res) => {
-				if (res.error) {
-					throwError(res.error, res.error.message);
-				}
-				notes = Array.from(
-					res.data?.map((note: any) => decryptNote(note, key)) || []
-				);
-				if (filterKey) {
-					notes = notes.filter(
-						(note) =>
-							note.title.includes(filterKey) ||
-							note.content.includes(filterKey)
-					);
-				}
-			});
-		return notes;
-	} catch (e) {
-		throwError(
-			e,
-			"Failed to get notes from Fleeting Notes - Check your credentials"
-		);
-	}
-	return notes;
-};
-
-interface SupabaseNote {
-	id: string;
-	title: string;
-	content: string;
-	source: string;
-	created_at: string;
-	modified_at: string;
-	deleted: boolean;
-	shared: boolean;
-	encrypted: boolean;
-	_partition: string;
-}
-
-export const updateNotesSupabase = async ({
-	firebaseId,
-	supabaseId,
-	key,
-	notes,
-}: {
-	firebaseId: string;
-	supabaseId: string;
-	key: string;
-	notes: Array<any>;
-}) => {
-	try {
-		let supabaseNotes: SupabaseNote[] = [];
-		let noteIds = new Set(notes.map((note) => note.id));
-		// get all fields of the note
-    const res = await supabase
-      .from("notes")
-      .select()
-      .in("_partition", [firebaseId, supabaseId])
-      .eq("deleted", false)
-
-    if (res.error) {
-      throwError(res.error, res.error.message);
-    }
-    supabaseNotes = res.data;
-
-    // only take notes that are modified after note from db & note exists
-    notes = notes.filter((note) => {
-      let supabaseNote = res.data.find(
-        (supabaseNote: any) => 
-        supabaseNote.id === note.id && noteIds.has(supabaseNote.id)
-      );
-      return (supabaseNote) ? true : false;
-    });
-
-    // merge possibly updated fields
-    notes = notes.map((note) => {
-      let supabaseNote = supabaseNotes?.find(
-        (supabaseNote: any) => supabaseNote.id === note.id
-      );
-      var newNote = {
-        ...supabaseNote,
-        title: note.title || supabaseNote.title,
-        content: note.content || supabaseNote.content,
-        source: note.source || supabaseNote.source,
-        modified_at: new Date().toISOString(),
-        deleted: note.deleted || supabaseNote.deleted,
-      };
-      return (supabaseNote.encrypted) ? encryptNote(newNote, key) : newNote;
-    });
-
-    if (notes.length > 0) {
-      const { error } = await supabase
-        .from("notes")
-        .upsert(notes, {
-          onConflict: "id",
-        })
-      if (error) {
-        throwError(error, error.message);
-      }
-    }
-	} catch (e) {
-		throwError(
-			e,
-			"Failed to update notes in Fleeting Notes"
-		);
-	}
-};
 
 export const decryptNote = (note: any, key: string) => {
 	if (note.encrypted) {
